Move the overlay's per-symbol contiguity case into the darts-quarantined file
`test_the_overlay_scopes_its_contiguity_check_per_symbol` calls
`_overlay_fold_temporal_features`, which imports darts and reaches
pytorch_lightning. `test-unit` installs neither, so the case failed collection
there with ModuleNotFoundError while its thirteen siblings - which exercise the
polars path and import only polars - passed. That is what turned test-unit red
on this branch.

`tests/test_darts_temporal.py` already exists for exactly this: it imports the
same private helper at module scope and is one of the four files test.yml names
as bringing darts, so it runs in `test-unit-image` against the reader's Docker
image. Moving the case there leaves the file it came from importing nothing
heavier than polars, which is what its remaining twelve cases are about.

No assertion changed. The local `_dates(100)` helper travels with it as
`_daily_dates`, and the two function-local imports drop because the destination
already has both at module scope.

test_uncovered_temporal_rows_are_dropped.py: 12 passed.
test_darts_temporal.py: 2 passed.
