+++ tests/test_darts_temporal.py
@@ -1,3 +1,5 @@
+import datetime as dt
+
 import pandas as pd
 
 from case_studies.utils.darts_forecasting import _overlay_fold_temporal_features
@@ -42,3 +44,54 @@ def test_darts_fold_temporal_overlay_uses_requested_fold_values():
 
     assert fold["garch_cond_vol"].to_list() == [1.1, 1.2]
     assert fold["other"].to_list() == [1.0, 2.0]
+
+
+def _daily_dates(n: int) -> list[dt.date]:
+    base = dt.date(2020, 1, 1)
+    return [base + dt.timedelta(days=i) for i in range(n)]
+
+
+def test_the_overlay_scopes_its_contiguity_check_per_symbol():
+    """Artifact coverage is ragged across symbols; the check must not read that as a hole.
+
+    etfs' fold 0 has six distinct artifact first-dates across its 100 symbols. Checked globally,
+    one symbol's legitimate leading trim falls strictly inside another symbol's date range and
+    reads as an interior hole, so a fold that is fine is refused. This is what passing
+    `entity_col` through every `_overlay_fold_temporal_features` call site buys.
+    """
+    dates = pd.to_datetime(_daily_dates(100))
+    dataset = pd.DataFrame(
+        {
+            "timestamp": [*dates, *dates],
+            "symbol": ["A"] * len(dates) + ["B"] * len(dates),
+            "garch_cond_vol": [-1.0] * (2 * len(dates)),
+        }
+    )
+    # A is covered from index 4, B from index 9 - a trim for each, no hole in either.
+    covered = [(s, d) for s, cut in (("A", 4), ("B", 9)) for d in dates[cut:]]
+    temporal = pd.DataFrame(
+        {
+            "timestamp": [d for _, d in covered],
+            "symbol": [s for s, _ in covered],
+            "fold": [0] * len(covered),
+            "garch_cond_vol": [0.5] * len(covered),
+        }
+    )
+
+    fold = _overlay_fold_temporal_features(
+        dataset,
+        {
+            "fold": 0,
+            "train_start": dates[0],
+            "train_end": dates[-1],
+            "val_end": dates[-1],
+        },
+        "timestamp",
+        temporal,
+        ["timestamp", "symbol"],
+        ["garch_cond_vol"],
+        "symbol",
+    )
+
+    assert (fold["symbol"] == "A").sum() == 96
+    assert (fold["symbol"] == "B").sum() == 91
