+++ tests/test_uncovered_temporal_rows_are_dropped.py
@@ -278,53 +278,3 @@ def test_a_trim_within_the_excuse_is_still_allowed():
         what="fold 0 train",
     )
     assert kept.height == 190
-
-
-def test_the_overlay_scopes_its_contiguity_check_per_symbol():
-    """Artifact coverage is ragged across symbols; the check must not read that as a hole.
-
-    etfs' fold 0 has six distinct artifact first-dates across its 100 symbols. Checked globally,
-    one symbol's legitimate leading trim falls strictly inside another symbol's date range and
-    reads as an interior hole, so a fold that is fine is refused. This is what passing
-    `entity_col` through every `_overlay_fold_temporal_features` call site buys.
-    """
-    import pandas as pd
-
-    from case_studies.utils.darts_forecasting import _overlay_fold_temporal_features
-
-    dates = pd.to_datetime(_dates(100))
-    dataset = pd.DataFrame(
-        {
-            "timestamp": [*dates, *dates],
-            "symbol": ["A"] * len(dates) + ["B"] * len(dates),
-            "garch_cond_vol": [-1.0] * (2 * len(dates)),
-        }
-    )
-    # A is covered from index 4, B from index 9 - a trim for each, no hole in either.
-    covered = [(s, d) for s, cut in (("A", 4), ("B", 9)) for d in dates[cut:]]
-    temporal = pd.DataFrame(
-        {
-            "timestamp": [d for _, d in covered],
-            "symbol": [s for s, _ in covered],
-            "fold": [0] * len(covered),
-            "garch_cond_vol": [0.5] * len(covered),
-        }
-    )
-
-    fold = _overlay_fold_temporal_features(
-        dataset,
-        {
-            "fold": 0,
-            "train_start": dates[0],
-            "train_end": dates[-1],
-            "val_end": dates[-1],
-        },
-        "timestamp",
-        temporal,
-        ["timestamp", "symbol"],
-        ["garch_cond_vol"],
-        "symbol",
-    )
-
-    assert (fold["symbol"] == "A").sum() == 96
-    assert (fold["symbol"] == "B").sum() == 91
